add: user now can see/write only the documents he own

+++ app/functions/remove.js
@@ -2,10 +2,10 @@ const { getDbConnection } = require('../db')
 const { ENTITY_NAME } = require('../constants')
 
 const remove = async (input) => {
-  const { id } = input;
+  const { id, user_id } = input;
   const db = await getDbConnection()
-  const query = `DELETE FROM ${ENTITY_NAME} WHERE id = $1 RETURNING id`;
-  const result = await db.query(query, [id]);
+  const query = `DELETE FROM ${ENTITY_NAME} WHERE id = $1 AND user_id = $2 RETURNING id`;
+  const result = await db.query(query, [id, user_id]);
   await db.end()
   return result.rowCount > 0;
 }
